fix: past deadlines message, eager load submission status on projects
