+++ frontend/package.json
@@ -19,6 +19,7 @@
         "@vuelidate/validators": "^2.0.4",
         "@vueuse/core": "^10.9.0",
         "axios": "^1.6.8",
+        "axios-cache-interceptor": "^1.5.2",
         "js-cookie": "^3.0.5",
         "moment": "^2.30.1",
         "msw": "^2.2.13",
