+++ frontend/src/types/Group.ts
@@ -1,4 +1,4 @@
-import { type Project } from './Project.ts';
+import { Project } from './Project.ts';
 import { type Student } from './users/Student.ts';
 import { type Submission } from './submission/Submission.ts';
 
@@ -6,10 +6,11 @@ export class Group {
     constructor(
         public id: string,
         public score: number = -1,
-        public project: Project | null = null,
+        public project: Project,
         public students: Student[] | null = null,
-        public submissions: Submission[] | null = null,
-    ) {}
+        public submissions: Submission[] | null = null
+    ) {
+    }
 
     /**
      * Check if the group is locked.
@@ -41,9 +42,14 @@ export class Group {
      * @param group
      */
     static fromJSON(group: Group): Group {
-        return new Group(group.id, group.score);
+        return new Group(group.id, group.score, Project.fromJSON(group.project));
     }
 
+    /**
+     * Convert a group object to a group instance.
+     *
+     * @param group
+     */
     static fromJSONFullObject(group: Group): Group {
         return new Group(group.id, group.score, group.project, group.students, group.submissions);
     }
