+++ frontend/src/types/Project.ts
@@ -4,6 +4,7 @@ import { type ExtraCheck } from './ExtraCheck.ts';
 import { type Group } from './Group.ts';
 import { type StructureCheck } from './StructureCheck.ts';
 import { type Submission } from './submission/Submission.ts';
+import { SubmissionStatus } from '@/types/SubmisionStatus.ts';
 
 export class Project {
     constructor(
@@ -18,7 +19,8 @@ export class Project {
         public max_score: number,
         public score_visible: boolean,
         public group_size: number,
-        public course: Course | null = null,
+        public course: Course,
+        public status: SubmissionStatus,
         public structure_file: File | null = null,
         public structureChecks: StructureCheck[] | null = null,
         public extra_checks: ExtraCheck[] | null = null,
@@ -100,12 +102,6 @@ export class Project {
      * @param project
      */
     static fromJSON(project: Project): Project {
-        let course = null;
-
-        if (project.course !== null && project.course !== undefined) {
-            course = Course.fromJSON(project.course);
-        }
-
         return new Project(
             project.id,
             project.name,
@@ -118,7 +114,8 @@ export class Project {
             project.max_score,
             project.score_visible,
             project.group_size,
-            course,
+            Course.fromJSON(project.course),
+            SubmissionStatus.fromJSON(project.status),
         );
     }
 }
